Dialect
A Dialect provides some convenience methods to convert values such as
type names from MOF to specific languages. Some methods set to
uppercase or to lowercase the initial letter of a string.
The Java class is a dialect. It provides type name conversions from Mof to Java.
Dialect and Java class are in the ELEvaluator subproject.
The Velocity action uses the dialect provided by its "dialect"
parameter. The default value is the Java dialect.